requirements.txt
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.3
packaging==24.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.1.0
six==1.16.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==2.0.1
